requirements.txt
streamlit
pandas
numpy
scikit-learn
matplotlib
reportlab
bcrypt
psycopg2-binary
plotly